아이템 별점, 리뷰 수정

+++ shop-master/src/main/java/com/shop/dto/BreadItemDto.java
@@ -29,10 +29,10 @@ public class BreadItemDto {
 
     private double grade;
     
-    private double commentsCount;
+    private int commentsCount;
     
     @QueryProjection
-    public BreadItemDto(Long id, String itemNm, String itemDetail, String imgUrl,int ori_price ,double price,double percent, ItemType itemType, String itemSummary, double grade, double commentsCount){
+    public BreadItemDto(Long id, String itemNm, String itemDetail, String imgUrl,int ori_price ,double price,double percent, ItemType itemType, String itemSummary, double grade, int commentsCount){
         this.id = id;
         this.itemNm = itemNm;
         this.itemDetail = itemDetail;
