+++ shop-master/src/main/java/com/shop/dto/ItemFormDto.java
@@ -52,7 +52,7 @@ public class ItemFormDto {
     
     private double grade;
     
-    private double commentsCount;
+    private int commentsCount;
     
     // 댓글
     private List<ItemCommentResponseDto> itemComments = new ArrayList<>();
